chore: clean test page
